Added checks for update and delete routes

+++ src/modules/users/services/DeleteUserService.ts
@@ -10,13 +10,16 @@ export class DeleteUserService {
     this.userRepository = getRepository(User);
   }
 
-  public async execute(id: number | string): Promise<void> {
+  public async execute(id: number | string, userId: string): Promise<void> {
     if (typeof id === 'string') {
       id = parseInt(id);
     }
     const user = await this.userRepository.findOne({ id });
     if (!user) {
       throw new AppError('User not found.', 404);
+    }
+    if (user.id !== parseInt(userId)) {
+      throw new AppError('You cannot delete other users.', 401);
     }
     await this.userRepository.delete(id);
   }
